remove everything from primitives crate root
